+++ packages/openai-compatible/src/openai-compatible-completion-language-model.test.ts
@@ -111,7 +111,6 @@ describe('doGenerate', () => {
     prepareJsonResponse({ content: 'Hello, World!' });
 
     const { content } = await model.doGenerate({
-      inputFormat: 'prompt',
       prompt: TEST_PROMPT,
     });
 
@@ -132,7 +131,6 @@ describe('doGenerate', () => {
     });
 
     const { usage } = await model.doGenerate({
-      inputFormat: 'prompt',
       prompt: TEST_PROMPT,
     });
 
@@ -143,7 +141,6 @@ describe('doGenerate', () => {
     prepareJsonResponse({});
 
     const { request } = await model.doGenerate({
-      inputFormat: 'prompt',
       prompt: TEST_PROMPT,
     });
 
@@ -156,9 +153,16 @@ describe('doGenerate', () => {
           "max_tokens": undefined,
           "model": "gpt-3.5-turbo-instruct",
           "presence_penalty": undefined,
-          "prompt": "Hello",
+          "prompt": "user:
+      Hello
+
+      assistant:
+      ",
           "seed": undefined,
-          "stop": undefined,
+          "stop": [
+            "
+      user:",
+          ],
           "suffix": undefined,
           "temperature": undefined,
           "top_p": undefined,
@@ -176,7 +180,6 @@ describe('doGenerate', () => {
     });
 
     const { response } = await model.doGenerate({
-      inputFormat: 'prompt',
       prompt: TEST_PROMPT,
     });
 
@@ -220,7 +223,6 @@ describe('doGenerate', () => {
     const { finishReason } = await provider
       .completionModel('gpt-3.5-turbo-instruct')
       .doGenerate({
-        inputFormat: 'prompt',
         prompt: TEST_PROMPT,
       });
 
@@ -236,7 +238,6 @@ describe('doGenerate', () => {
     const { finishReason } = await provider
       .completionModel('gpt-3.5-turbo-instruct')
       .doGenerate({
-        inputFormat: 'prompt',
         prompt: TEST_PROMPT,
       });
 
@@ -249,7 +250,6 @@ describe('doGenerate', () => {
     });
 
     const { response } = await model.doGenerate({
-      inputFormat: 'prompt',
       prompt: TEST_PROMPT,
     });
 
@@ -267,14 +267,23 @@ describe('doGenerate', () => {
     prepareJsonResponse({ content: '' });
 
     await model.doGenerate({
-      inputFormat: 'prompt',
       prompt: TEST_PROMPT,
     });
 
-    expect(await server.calls[0].requestBody).toStrictEqual({
-      model: 'gpt-3.5-turbo-instruct',
-      prompt: 'Hello',
-    });
+    expect(await server.calls[0].requestBody).toMatchInlineSnapshot(`
+      {
+        "model": "gpt-3.5-turbo-instruct",
+        "prompt": "user:
+      Hello
+
+      assistant:
+      ",
+        "stop": [
+          "
+      user:",
+        ],
+      }
+    `);
   });
 
   it('should pass headers', async () => {
@@ -290,7 +299,6 @@ describe('doGenerate', () => {
     });
 
     await provider.completionModel('gpt-3.5-turbo-instruct').doGenerate({
-      inputFormat: 'prompt',
       prompt: TEST_PROMPT,
       headers: {
         'Custom-Request-Header': 'request-header-value',
@@ -309,7 +317,6 @@ describe('doGenerate', () => {
     prepareJsonResponse({ content: '' });
 
     await provider.completionModel('gpt-3.5-turbo-instruct').doGenerate({
-      inputFormat: 'prompt',
       prompt: TEST_PROMPT,
       providerOptions: {
         'test-provider': {
@@ -318,18 +325,27 @@ describe('doGenerate', () => {
       },
     });
 
-    expect(await server.calls[0].requestBody).toStrictEqual({
-      model: 'gpt-3.5-turbo-instruct',
-      prompt: 'Hello',
-      someCustomOption: 'test-value',
-    });
+    expect(await server.calls[0].requestBody).toMatchInlineSnapshot(`
+      {
+        "model": "gpt-3.5-turbo-instruct",
+        "prompt": "user:
+      Hello
+
+      assistant:
+      ",
+        "someCustomOption": "test-value",
+        "stop": [
+          "
+      user:",
+        ],
+      }
+    `);
   });
 
   it('should not include provider-specific options for different provider', async () => {
     prepareJsonResponse({ content: '' });
 
     await provider.completionModel('gpt-3.5-turbo-instruct').doGenerate({
-      inputFormat: 'prompt',
       prompt: TEST_PROMPT,
       providerOptions: {
         notThisProviderName: {
@@ -338,10 +354,20 @@ describe('doGenerate', () => {
       },
     });
 
-    expect(await server.calls[0].requestBody).toStrictEqual({
-      model: 'gpt-3.5-turbo-instruct',
-      prompt: 'Hello',
-    });
+    expect(await server.calls[0].requestBody).toMatchInlineSnapshot(`
+      {
+        "model": "gpt-3.5-turbo-instruct",
+        "prompt": "user:
+      Hello
+
+      assistant:
+      ",
+        "stop": [
+          "
+      user:",
+        ],
+      }
+    `);
   });
 });
 
@@ -398,7 +424,6 @@ describe('doStream', () => {
     });
 
     const { stream } = await model.doStream({
-      inputFormat: 'prompt',
       prompt: TEST_PROMPT,
     });
 
@@ -454,7 +479,6 @@ describe('doStream', () => {
     };
 
     const { stream } = await model.doStream({
-      inputFormat: 'prompt',
       prompt: TEST_PROMPT,
     });
 
@@ -494,7 +518,6 @@ describe('doStream', () => {
       };
 
       const { stream } = await model.doStream({
-        inputFormat: 'prompt',
         prompt: TEST_PROMPT,
       });
 
@@ -526,7 +549,6 @@ describe('doStream', () => {
     prepareStreamResponse({ content: [] });
 
     const { request } = await model.doStream({
-      inputFormat: 'prompt',
       prompt: TEST_PROMPT,
     });
 
@@ -539,9 +561,16 @@ describe('doStream', () => {
           "max_tokens": undefined,
           "model": "gpt-3.5-turbo-instruct",
           "presence_penalty": undefined,
-          "prompt": "Hello",
+          "prompt": "user:
+      Hello
+
+      assistant:
+      ",
           "seed": undefined,
-          "stop": undefined,
+          "stop": [
+            "
+      user:",
+          ],
           "stream": true,
           "stream_options": undefined,
           "suffix": undefined,
@@ -559,7 +588,6 @@ describe('doStream', () => {
     });
 
     const { response } = await model.doStream({
-      inputFormat: 'prompt',
       prompt: TEST_PROMPT,
     });
 
@@ -578,16 +606,24 @@ describe('doStream', () => {
     prepareStreamResponse({ content: [] });
 
     await model.doStream({
-      inputFormat: 'prompt',
       prompt: TEST_PROMPT,
     });
 
-    expect(await server.calls[0].requestBody).toStrictEqual({
-      stream: true,
-      // stream_options: { include_usage: true },
-      model: 'gpt-3.5-turbo-instruct',
-      prompt: 'Hello',
-    });
+    expect(await server.calls[0].requestBody).toMatchInlineSnapshot(`
+      {
+        "model": "gpt-3.5-turbo-instruct",
+        "prompt": "user:
+      Hello
+
+      assistant:
+      ",
+        "stop": [
+          "
+      user:",
+        ],
+        "stream": true,
+      }
+    `);
   });
 
   it('should pass headers', async () => {
@@ -603,7 +639,6 @@ describe('doStream', () => {
     });
 
     await provider.completionModel('gpt-3.5-turbo-instruct').doStream({
-      inputFormat: 'prompt',
       prompt: TEST_PROMPT,
       headers: {
         'Custom-Request-Header': 'request-header-value',
@@ -622,7 +657,6 @@ describe('doStream', () => {
     prepareStreamResponse({ content: [] });
 
     await model.doStream({
-      inputFormat: 'prompt',
       providerOptions: {
         'test-provider': {
           someCustomOption: 'test-value',
@@ -631,19 +665,28 @@ describe('doStream', () => {
       prompt: TEST_PROMPT,
     });
 
-    expect(await server.calls[0].requestBody).toStrictEqual({
-      stream: true,
-      model: 'gpt-3.5-turbo-instruct',
-      prompt: 'Hello',
-      someCustomOption: 'test-value',
-    });
+    expect(await server.calls[0].requestBody).toMatchInlineSnapshot(`
+      {
+        "model": "gpt-3.5-turbo-instruct",
+        "prompt": "user:
+      Hello
+
+      assistant:
+      ",
+        "someCustomOption": "test-value",
+        "stop": [
+          "
+      user:",
+        ],
+        "stream": true,
+      }
+    `);
   });
 
   it('should not include provider-specific options for different provider', async () => {
     prepareStreamResponse({ content: [] });
 
     await model.doStream({
-      inputFormat: 'prompt',
       providerOptions: {
         notThisProviderName: {
           someCustomOption: 'test-value',
@@ -652,10 +695,20 @@ describe('doStream', () => {
       prompt: TEST_PROMPT,
     });
 
-    expect(await server.calls[0].requestBody).toStrictEqual({
-      stream: true,
-      model: 'gpt-3.5-turbo-instruct',
-      prompt: 'Hello',
-    });
+    expect(await server.calls[0].requestBody).toMatchInlineSnapshot(`
+      {
+        "model": "gpt-3.5-turbo-instruct",
+        "prompt": "user:
+      Hello
+
+      assistant:
+      ",
+        "stop": [
+          "
+      user:",
+        ],
+        "stream": true,
+      }
+    `);
   });
 });