+++ packages/provider/src/embedding-model/v2/embedding-model-v2.ts
@@ -31,13 +31,18 @@ Provider-specific model ID for logging purposes.
 
   /**
 Limit of how many embeddings can be generated in a single API call.
+
+Use Infinity for models that do not have a limit.
    */
-  readonly maxEmbeddingsPerCall: number | undefined;
+  readonly maxEmbeddingsPerCall:
+    | PromiseLike<number | undefined>
+    | number
+    | undefined;
 
   /**
 True if the model can handle multiple embedding calls in parallel.
    */
-  readonly supportsParallelCalls: boolean;
+  readonly supportsParallelCalls: PromiseLike<boolean> | boolean;
 
   /**
 Generates a list of embeddings for the given input text.