+++ packages/google/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @ai-sdk/google
 
+## 2.0.0-canary.13
+
+### Patch Changes
+
+- 7378473: chore(providers/google): switch to providerOptions
+- Updated dependencies [9bf7291]
+- Updated dependencies [4617fab]
+- Updated dependencies [e030615]
+  - @ai-sdk/provider@2.0.0-canary.11
+  - @ai-sdk/provider-utils@3.0.0-canary.12
+
 ## 2.0.0-canary.12
 
 ### Patch Changes